TakenState moved to helpers
